pyproject.toml
[project]
name = "hermes-vcc"
version = "0.1.0"
description = "Lossless conversation archiving and VCC-enhanced compression for Hermes agent"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]
readme = "README.md"
authors = [
    {name = "Nous Research"},
]
keywords = ["hermes", "vcc", "context-compression", "conversation-archiving", "ai-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://github.com/nousresearch/hermes-vcc"
Repository = "https://github.com/nousresearch/hermes-vcc"
Documentation = "https://github.com/nousresearch/hermes-vcc/tree/main/docs"
Issues = "https://github.com/nousresearch/hermes-vcc/issues"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-xdist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: requires LLM calls"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hermes_vcc"]

[tool.hatch.build.targets.sdist]
include = [
    "hermes_vcc/",
    "vendor/",
    "tests/",
    "docs/",
    "examples/",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]
